Reorganized files
Se organizan los archivos en:
- Capa de Dominio: entidades que resuelven lógica de negocio
- Capa de Aplicación: clases que colaboran a resolver la lógica del dominio.
