+++ src/VillaPeruana/Products/Domain/Entities/PiscoPeruano.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App;
+namespace App\VillaPeruana\Products\Domain;
 
 final class PiscoPeruano extends Product
 {
