+++ src/VillaPeruana/Products/Domain/Entities/TicketVipAlConciertoDePickFloid.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App;
+namespace App\VillaPeruana\Products\Domain;
 
 final class TicketVipAlConciertoDePickFloid extends Product
 {
