+++ src/VillaPeruana/Products/Domain/Entities/TumiDeOroMoche.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App;
+namespace App\VillaPeruana\Products\Domain;
 
 final class TumiDeOroMoche extends Product
 {
